refactor fetch api calls to a separate file
